+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add a dry-run option to the CLI update command and add logging to the update process
 - Refactor MongoLite + add support for more MongoDB aggregation operators
 - Fix the possibility to delete files outside of Cockpit as super admin
+- Fix Async code generation
 
 ## 2.13.3 (2026-01-09)